+++ NEWS.md
@@ -1,6 +1,9 @@
 ## v3.0.4 2023-11-30
 ### Bug fixes
 * Fix shared filter for subjects/contributors ([MSEARCH-639](https://issues.folio.org/browse/MSEARCH-639))
+* Return result only for desired cn type on browse ([MSEARCH-605](https://issues.folio.org/browse/MSEARCH-605))
+
+---
 
 ## v3.0.3 2023-11-15
 ### Tech Dept